+++ packages/demo/src/index.tsx
@@ -1,12 +1,11 @@
 import {
+  createDocumentIdentifier,
+  Document,
+  Editor,
+  EditorProps,
   EditorContext,
   EditorUtilsContext
-} from '@splish-me/editor-core-contexts'
-import { Editor, EditorProps } from '@splish-me/editor-core'
-import {
-  createDocumentIdentifier,
-  Document
-} from '@splish-me/editor-core-document'
+} from '@splish-me/editor'
 import { ModeToolbar } from '@splish-me/editor-ui-mode-toolbar'
 import { AddSidebar } from '@splish-me/editor-ui-add-sidebar'
 import { Sidebar } from '@splish-me/editor-ui-sidebar'