+++ packages/editor/src/editor/document-editor/index.tsx
@@ -1,16 +1,10 @@
-import { createDocumentIdentifier } from '@splish-me/editor-core-document'
-import {
-  DocumentIdentifier,
-  DocumentProps,
-  Plugin,
-  SerializedDocument,
-  State
-} from '@splish-me/editor-core-types'
 import { editable } from 'ory-editor-core/lib/selector/editable'
 import { createEmptyState, Editor } from 'ory-editor-core'
 import * as React from 'react'
 import * as R from 'ramda'
 
+import { createDocumentIdentifier } from '../../document'
+import { DocumentIdentifier, Plugin, State } from '../../types'
 import { DocumentInner } from './inner'
 
 export class DocumentEditor extends React.Component<DocumentEditorProps> {