+++ scripts/changelog.ts
@@ -16,9 +16,16 @@ async function exec(): Promise<void> {
       date: '2019-01-14'
     },
     {
-      breakingChanges: ["Don't provide `module` entry anymore"],
+      breakingChanges: [
+        '@splish-me/editor replaces @splish-me/editor-core, @splish-me/editor-core-contexts, @splish-me/editor-core-document and @splish-me/editor-core-types',
+        '@splish-me/editor-html-renderer replaces @splish-me/editor-core-html-renderer',
+        '@splish-me/editor-renderer replaces @splish-me/editor-core-renderer',
+        "Don't provide `module` entry anymore"
+      ],
       added: ['Pass plugins as dictionary'],
-      deprecated: ['Deprecate passing plugins as array'],
+      deprecated: [
+        'Passing plugins as array is deprecated and will be removed in the next minor version. Pass plugins as dictionary instead.'
+      ],
       fixed: [
         'Serialize nested documents correctly',
         'Handle edge cases of undo/redo correctly'